Adding Vagrantfile to launch an AWS EC2 instance and run Docker provisioner
This patch adds a Vagrantfile which launches an AWS EC2 instance using aws
as the provider. It expectes few of the run time parameters from the ENV
variables. These things will be explained in README

It also adds a Dockerfile which will be synced to the remote guest via rsync.
So the instance where you are running should have rsync inatalled.

To test the things, Vagrantfile used docker provisioner to build an apache
image and then running it in port 80 of the ec2 instance.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>